+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Types of changes: `Added`, `Changed`, `Deprecate`, `Removed`, `Fixed`, `Secruity
 
 ## [Unreleased]
 
+### Added
+- Added tests for SQLite
+
 ### Changed
 - Changed the test case implementation to fix issue that class name must be unique
 - Changed function signatures for PHP 8.4 compatibility